+++ apps/demo/src/router/table.ts
@@ -1,0 +1,23 @@
+const Table = () => import("@/views/Table/Index.vue");
+const BasicTable = () => import("@/views/Table/BasicTable.vue");
+
+
+const routes = [
+  {
+    path: "/table",
+    children: [
+      {
+        component: Table,
+        name: "table",
+        path: "",
+      },
+      {
+        component: BasicTable,
+        name: "basicTable",
+        path: "basic",
+      },
+    ],
+  },
+];
+
+export default routes;