+++ apps/demo/src/views/Table/Index.vue
@@ -1,0 +1,13 @@
+<template>
+  <TablePage :title="$t('table.title')" />
+</template>
+
+<script lang="ts">
+export default {
+  name: "TableDemo",
+};
+</script>
+
+<script setup lang="ts">
+import TablePage from "./TablePage.vue";
+</script>