+++ packages/vue-layout/src/Layout.vue
@@ -1,19 +1,13 @@
 <template>
   <component :is="layoutComponent">
     <slot></slot>
+    <template v-for="name in slotNames" #[name]>
+      <slot :name="name"></slot>
+    </template>
   </component>
 </template>
 
 <script lang="ts">
-/* eslint-disable-next-line import/order */
-import {
-  defineAsyncComponent,
-  computed,
-  markRaw,
-  shallowRef,
-  watch,
-} from "vue";
-
 const NullLayout = defineAsyncComponent({
   loader: () => import("./layouts/NullLayout.vue"),
   timeout: 3000,
@@ -25,6 +19,14 @@ export default {
 </script>
 
 <script setup lang="ts">
+import {
+  defineAsyncComponent,
+  computed,
+  markRaw,
+  shallowRef,
+  useSlots,
+  watch,
+} from "vue";
 import { useRoute } from "vue-router";
 
 import type { LayoutType } from "./types";
@@ -37,6 +39,9 @@ const props = defineProps({
 });
 
 const route = useRoute();
+
+const slots = useSlots();
+const slotNames = Object.keys(slots) as string[];
 
 let layout = shallowRef(props.defaultLayout);