+++ packages/vue-tanstack-table/src/index.ts
@@ -1,5 +1,8 @@
-import "./assets/css/index.css";
-
+import Table from "./components/Table.vue";
+import TableBody from "./components/TableBody.vue";
+import TableHeader from "./components/TableHeader.vue";
 import TanstackTable from "./components/TanstackTable.vue";
 
-export { TanstackTable };
+export { Table, TableBody, TableHeader, TanstackTable };
+
+export type { TableColumnDefinition } from "./types";